requirements.txt
absl-py==0.8.1
astor==0.8.0
astroid==2.2.5
astunparse==1.6.3
autopep8==1.5
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
gast==0.3.3
google-auth==1.15.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.9
isort==4.3.16
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==3.1.1
matplotlib==3.2.1
mccabe==0.6.1
nltk==3.4.5
numpy==1.18.4
oauthlib==3.1.0
opt-einsum==3.2.1
Pillow==7.1.2
protobuf==3.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pylint==2.3.1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rope==0.14.0
rsa==4.0
scipy==1.4.1
six==1.13.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
typed-ast==1.3.1
urllib3==1.25.9
virtualenv==16.0.0
Werkzeug==0.16.0
wrapt==1.11.2
